pyproject.toml
[project]
name = "learning-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "langchain>=1.2.6",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=7.0.0",
]
